updated testcases

+++ src/main/java/com/hexaware/automobile/insurancesystem/service/DocumentServiceImp.java
@@ -8,19 +8,32 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.hexaware.automobile.insurancesystem.dto.DocumentDto;
 import com.hexaware.automobile.insurancesystem.entities.Document;
+import com.hexaware.automobile.insurancesystem.entities.Proposal;
+import com.hexaware.automobile.insurancesystem.exception.ClaimNotFoundException;
 import com.hexaware.automobile.insurancesystem.exception.DocumentNotFoundException;
 import com.hexaware.automobile.insurancesystem.repository.DocumentRepository;
+import com.hexaware.automobile.insurancesystem.repository.ProposalRepository;
 @Service
 public class DocumentServiceImp implements IDocumentService{
 
 	@Autowired
 	DocumentRepository repo;
+	@Autowired
+	  ProposalRepository repo1;
 	
 	@Override
-	public Document addDocument(Document document) {
-		
-		return repo.save(document);
+	public Document addDocument(DocumentDto dto) {
+		Document document = new Document();
+        document.setDocId(dto.getDocId());
+        document.setDocType(dto.getDocType());
+
+        Proposal proposal = repo1.findById(dto.getProposalId())
+                .orElseThrow(() -> new RuntimeException("Proposal not found"));
+        document.setProposal(proposal);
+
+        return repo.save(document);
 	}
 
 	@Override
@@ -45,9 +58,17 @@ public class DocumentServiceImp implements IDocumentService{
 
 	@Override
 	public String deleteDocumentById(int docId) {
-		repo.deleteById(docId);
-		
-		return "Recoed deleted successfully";
+		if (!repo.existsById(docId)) {
+            throw new ClaimNotFoundException("Cannot delete - claim not found");
+        }
+        repo.deleteById(docId);
+        return "Record deleted successfully";
+
+	}
+
+	@Override
+	public List<Document> getDocumentsByProposalId(int proposalId) {
+		 return repo.findByProposal_ProposalId(proposalId);
 	}
 
 }
